requirements.txt
streamlit>=1.40.0
faster-whisper>=1.1.0
pyannote.audio>=3.3.0
pydub>=0.25.1
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
